refactor out useLoadWebChat and add unit test
