+++ src/applications/virtual-agent/tests/hooks/useLoadWebChat.unit.spec.js
@@ -1,0 +1,37 @@
+import sinon from 'sinon';
+import { expect } from 'chai';
+
+import { renderHook } from '@testing-library/react-hooks';
+import * as UseFeatureToggleModule from '~/platform/utilities/feature-toggles/useFeatureToggle/';
+import useLoadWebChat from '../../hooks/useLoadWebChat';
+
+describe('useLoadWebChat', () => {
+  let sandbox;
+
+  beforeEach(() => {
+    sandbox = sinon.sandbox.create();
+  });
+
+  afterEach(() => {
+    sandbox.restore();
+  });
+
+  describe('useLoadWebChat', () => {
+    it('should load regular webchat virtualAgentUpgradeWebchat14158 feature toggle is false', () => {
+      const useToggleValueStub = sandbox.stub().returns(false);
+      sandbox.stub(UseFeatureToggleModule, 'useFeatureToggle').returns({
+        useToggleValue: useToggleValueStub,
+        TOGGLE_NAMES: {
+          virtualAgentUpgradeWebchat14158: 'virtualAgentUpgradeWebchat14158',
+        },
+      });
+
+      renderHook(() => useLoadWebChat());
+
+      const script = document.querySelector('script');
+      expect(script.src).to.equal(
+        'https://cdn.botframework.com/botframework-webchat/4.15.8/webchat.js',
+      );
+    });
+  });
+});
